parasite_words_clf/requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
filelock==3.11.0
gdown==4.7.1
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-dataclass==8.5.14
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
packaging==23.1
pydantic==1.10.7
PySocks==1.7.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
